pyproject.toml
[tool.poetry]
name = "wc-guide-backend"
version = "0.1.0"
description = ""
authors = ["murthy10 <samuel.kurath@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.3"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.8.6"
djangorestframework-gis = "^0.17"
django-cors-headers = "^3.7.0"
drf-yasg = "^1.20.0"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-django = "^4.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
